Move the resolve_execution_environment method to the mixin class
so that it can be used with AdHocCommands as well.
